fix(cache): Fix cache hit counting, add TTL expiration stats, and consolidate MySQL binlog tests
Fix incorrect cache hit counting and add TTL expiration tracking:

Cache fixes:
- Defer cache_hits++ until after decompression succeeds in both Lookup
  and LookupWithMetadata, so failed decompressions do not inflate stats
- Add null check for invalidation_mgr_ in InvalidationQueue::Enqueue
  before calling UnregisterCacheEntry to prevent null pointer dereference
- Extract RemoveEntryLocked helper in QueryCache to unify entry removal
  logic across LRU eviction, TTL expiration, and table clear paths

TTL expiration stats:
- Add ttl_expirations counter to CacheStatistics and CacheStatisticsSnapshot
- Add RemovalReason enum (kLRUEviction, kTTLExpired, kTableClear) to
  distinguish removal causes
- Expose ttl_expirations in INFO response (TCP and HTTP), Prometheus
  metrics (mygramdb_cache_ttl_expirations_total), and CACHE STATS command

Test coverage:
- Add StatsInvariantHitsPlusMissesEqualsTotal test verifying
  total_queries == cache_hits + cache_misses at all times
- Add TTLExpirationStats test verifying TTL removals increment
  ttl_expirations, not evictions

MySQL binlog test consolidation:
- Delete binlog_reader_bug_fixes_test.cpp and binlog_event_parser_bug_fixes_test.cpp
- Merge content into binlog_reader_core_test.cpp, binlog_reader_events_test.cpp,
  binlog_reader_multitable_test.cpp, and binlog_parsing_test.cpp
- Rename binlog_event_processor_bug_fixes_test.cpp to binlog_event_processor_test.cpp
- Extract shared fixtures into binlog_test_fixtures.h (BinlogReaderFixture,
  MakeDefaultTableConfig, MakeEvent helpers)
- Update CMakeLists.txt to remove three separate bug-fix test targets and
  register the consolidated binlog_event_processor_test target
