+++ src/server/handlers/cache_handler.cpp
@@ -77,6 +77,7 @@ std::string CacheHandler::HandleStats() {
   oss << "current_entries: " << stats.current_entries << "\r\n";
   oss << "current_memory_bytes: " << stats.current_memory_bytes << "\r\n";
   oss << "evictions: " << stats.evictions << "\r\n";
+  oss << "ttl_expirations: " << stats.ttl_expirations << "\r\n";
 
   // Invalidation statistics
   oss << "invalidations_immediate: " << stats.invalidations_immediate << "\r\n";
