+++ tests/mysql/CMakeLists.txt
@@ -216,48 +216,17 @@ if(USE_MYSQL AND MYSQL_FOUND)
     PROPERTIES RESOURCE_LOCK mysql_connection
   )
 
-  # Binlog reader bug fix tests (TDD)
-  add_executable(binlog_reader_bug_fixes_test
-    binlog_reader_bug_fixes_test.cpp
+  # Binlog event processor tests
+  add_executable(binlog_event_processor_test
+    binlog_event_processor_test.cpp
   )
 
-  target_link_libraries(binlog_reader_bug_fixes_test
+  target_link_libraries(binlog_event_processor_test
     mygramdb_server
     GTest::gtest_main
   )
 
-  gtest_discover_tests(binlog_reader_bug_fixes_test
-    DISCOVERY_MODE POST_BUILD
-    DISCOVERY_TIMEOUT 30
-    PROPERTIES RESOURCE_LOCK mysql_connection
-  )
-
-  # Binlog event parser bug fix tests (TDD)
-  add_executable(binlog_event_parser_bug_fixes_test
-    binlog_event_parser_bug_fixes_test.cpp
-  )
-
-  target_link_libraries(binlog_event_parser_bug_fixes_test
-    mygramdb_server
-    GTest::gtest_main
-  )
-
-  gtest_discover_tests(binlog_event_parser_bug_fixes_test
-    DISCOVERY_MODE POST_BUILD
-    DISCOVERY_TIMEOUT 30
-  )
-
-  # Binlog event processor bug fix tests (TDD)
-  add_executable(binlog_event_processor_bug_fixes_test
-    binlog_event_processor_bug_fixes_test.cpp
-  )
-
-  target_link_libraries(binlog_event_processor_bug_fixes_test
-    mygramdb_server
-    GTest::gtest_main
-  )
-
-  gtest_discover_tests(binlog_event_processor_bug_fixes_test
+  gtest_discover_tests(binlog_event_processor_test
     DISCOVERY_MODE POST_BUILD
     DISCOVERY_TIMEOUT 30
   )
