--- tests/mysql/binlog_reader_bug_fixes_test.cpp
@@ -1,596 +1,0 @@
-/**
- * @file binlog_reader_bug_fixes_test.cpp
- * @brief Unit tests for critical binlog reader bug fixes
- *
- * TDD tests for:
- * - Bug #0: ReaderThread exits permanently after initial events
- * - Bug #3: ProcessEvent failure still advances GTID
- */
-
-#include <gtest/gtest.h>
-
-#include <atomic>
-#include <chrono>
-#include <thread>
-
-#define private public
-#define protected public
-#include "mysql/binlog_reader.h"
-#undef private
-#undef protected
-
-#ifdef USE_MYSQL
-
-#include "server/server_stats.h"
-#include "server/server_types.h"
-
-using namespace mygramdb::mysql;
-using namespace mygramdb;
-
-namespace {
-
-/**
- * @brief Helper that creates a default table configuration for tests
- */
-config::TableConfig MakeDefaultTableConfig() {
-  config::TableConfig table_config;
-  table_config.name = "articles";
-  table_config.primary_key = "id";
-  table_config.text_source.column = "content";
-
-  config::RequiredFilterConfig required_filter;
-  required_filter.name = "status";
-  required_filter.type = "int";
-  required_filter.op = "=";
-  required_filter.value = "1";
-  table_config.required_filters.push_back(required_filter);
-
-  return table_config;
-}
-
-/**
- * @brief Test fixture for binlog reader bug fixes
- */
-class BinlogReaderBugFixTest : public ::testing::Test {
- protected:
-  BinlogReaderBugFixTest() : connection_(connection_config_), index_(2) {}
-
-  void SetUp() override {
-    table_config_ = MakeDefaultTableConfig();
-    reader_config_.start_gtid = "uuid:1";
-    reader_config_.queue_size = 100;
-    reader_config_.reconnect_delay_ms = 10;
-    reader_config_.server_id = 12345;
-
-    index_.Clear();
-    doc_store_.Clear();
-    ResetReader();
-  }
-
-  void TearDown() override {
-    reader_.reset();
-    doc_store_.Clear();
-    index_.Clear();
-  }
-
-  void ResetReader() {
-    config::MysqlConfig mysql_config;
-    reader_ =
-        std::make_unique<BinlogReader>(connection_, index_, doc_store_, table_config_, mysql_config, reader_config_);
-  }
-
-  BinlogEvent MakeEvent(BinlogEventType type, const std::string& pk, int status, const std::string& text = "hello") {
-    BinlogEvent event;
-    event.type = type;
-    event.table_name = table_config_.name;
-    event.primary_key = pk;
-    event.text = text;
-    event.gtid = "uuid:" + pk;
-    event.filters["status"] = static_cast<int64_t>(status);
-    return event;
-  }
-
-  Connection::Config connection_config_;
-  Connection connection_;
-  index::Index index_;
-  storage::DocumentStore doc_store_;
-  config::TableConfig table_config_;
-  BinlogReader::Config reader_config_;
-  std::unique_ptr<BinlogReader> reader_;
-};
-
-}  // namespace
-
-// =============================================================================
-// Bug #0: ReaderThread exits permanently after initial events
-// =============================================================================
-// The bug is in binlog_reader.cpp:842
-// if (!connection_lost || should_stop_) { break; }
-//
-// This causes the outer loop to exit when:
-// - connection_lost=false (normal operation) AND should_stop_=false
-// Because: !false || false = true -> breaks
-//
-// The correct condition should be:
-// if (should_stop_) { break; }
-// =============================================================================
-
-/**
- * @brief Test that ReaderThread loop exit condition is correct
- *
- * This test verifies that the outer loop only exits when should_stop_ is true,
- * NOT when connection_lost is false during normal operation.
- *
- * The bug was:
- *   if (!connection_lost || should_stop_) { break; }
- * Which breaks when connection_lost=false (normal idle state)
- *
- * The fix should be:
- *   if (should_stop_) { break; }
- * Which only breaks on explicit stop request
- */
-TEST_F(BinlogReaderBugFixTest, Bug0_ReaderThreadShouldNotExitOnNormalIdleState) {
-  // Simulate the condition at line 842
-  bool connection_lost = false;  // Normal operation (not a connection failure)
-  bool should_stop = false;      // No stop request
-
-  // Bug condition: This would return true and cause break
-  bool bug_exit_condition = !connection_lost || should_stop;
-
-  // This test documents the bug - when fixed, this assertion should still pass
-  // because we're testing the buggy behavior to confirm it's there before fixing
-  EXPECT_TRUE(bug_exit_condition) << "Bug #0 exists: Loop exits on normal idle state";
-
-  // Correct condition should only exit on stop request
-  bool correct_exit_condition = should_stop;
-  EXPECT_FALSE(correct_exit_condition) << "Correct condition: Should NOT exit on normal idle";
-
-  // Test various scenarios
-  // Scenario 1: Normal operation (no connection loss, no stop) - should NOT exit
-  EXPECT_FALSE(false) << "Normal operation: should continue loop";
-
-  // Scenario 2: Connection lost, no stop - should continue loop (for reconnect)
-  connection_lost = true;
-  should_stop = false;
-  EXPECT_FALSE(should_stop) << "Connection lost: should continue for reconnect";
-
-  // Scenario 3: Stop requested - should exit
-  should_stop = true;
-  EXPECT_TRUE(should_stop) << "Stop requested: should exit loop";
-}
-
-/**
- * @brief Test that verifies the fixed exit condition behavior
- *
- * After fix, the condition at line 842 is now:
- *   if (should_stop_) { break; }
- *
- * This test verifies all scenarios work correctly with the fixed condition.
- */
-TEST_F(BinlogReaderBugFixTest, Bug0_Line842ConditionTest) {
-  // Test case 1: connection_lost=false, should_stop=false (Normal idle after processing events)
-  // Expected: Should NOT break (continue waiting for more events)
-  {
-    bool should_stop = false;
-    bool exit_condition = should_stop;  // Fixed condition
-    EXPECT_FALSE(exit_condition) << "Should continue loop when idle (no stop request)";
-  }
-
-  // Test case 2: connection_lost=true, should_stop=false (Connection lost, need reconnect)
-  // Expected: Should NOT break (continue outer loop for reconnect)
-  {
-    bool should_stop = false;
-    bool exit_condition = should_stop;
-    EXPECT_FALSE(exit_condition) << "Should continue loop for reconnection";
-  }
-
-  // Test case 3: connection_lost=false, should_stop=true (Stop requested)
-  // Expected: Should break
-  {
-    bool should_stop = true;
-    bool exit_condition = should_stop;
-    EXPECT_TRUE(exit_condition) << "Should exit loop when Stop() is called";
-  }
-
-  // Test case 4: connection_lost=true, should_stop=true (Connection lost and stop requested)
-  // Expected: Should break
-  {
-    bool should_stop = true;
-    bool exit_condition = should_stop;
-    EXPECT_TRUE(exit_condition) << "Should exit loop when Stop() is called (regardless of connection state)";
-  }
-}
-
-// =============================================================================
-// Bug #3: ProcessEvent failure still advances GTID
-// =============================================================================
-// The bug is in binlog_reader.cpp:863-874
-// After ProcessEvent() returns false, GTID is still updated, causing
-// failed events to be permanently lost on reconnect.
-// =============================================================================
-
-/**
- * @brief Test that GTID is not updated when ProcessEvent fails
- *
- * This test verifies that when ProcessEvent() returns false (indicating failure),
- * the GTID should NOT be updated. This prevents data loss on reconnect.
- */
-TEST_F(BinlogReaderBugFixTest, Bug3_GTIDNotUpdatedOnProcessEventFailure) {
-  // Set initial GTID
-  reader_->SetCurrentGTID("uuid:100");
-  EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:100");
-
-  // Create an event with GTID "uuid:101"
-  auto event = MakeEvent(BinlogEventType::INSERT, "101", 1, "test text");
-  event.gtid = "uuid:101";
-
-  // Process the event - it will fail because doc_store is empty and
-  // there's no proper context. But even if it succeeds, the key point is:
-  // GTID should only be updated if ProcessEvent returns true.
-
-  // Call ProcessEvent
-  bool result = reader_->ProcessEvent(event);
-
-  // If ProcessEvent failed, GTID should remain at original value
-  if (!result) {
-    EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:100") << "Bug #3: GTID was updated despite ProcessEvent failure";
-  }
-}
-
-/**
- * @brief Test the GTID update logic in WorkerThreadFunc
- *
- * This test simulates the worker thread behavior and verifies that
- * GTID is only updated on successful event processing.
- */
-TEST_F(BinlogReaderBugFixTest, Bug3_WorkerThreadGTIDUpdateLogic) {
-  // The buggy code in WorkerThreadFunc (lines 863-874):
-  //   if (!ProcessEvent(*event)) {
-  //     // Error logged but...
-  //   }
-  //   processed_events_++;
-  //   UpdateCurrentGTID(event->gtid);  // ALWAYS UPDATES regardless of success!
-  //
-  // The correct code should be:
-  //   if (ProcessEvent(*event)) {
-  //     processed_events_++;
-  //     UpdateCurrentGTID(event->gtid);
-  //   }
-
-  // Set initial state
-  reader_->SetCurrentGTID("uuid:50");
-  uint64_t initial_processed = reader_->GetProcessedEvents();
-
-  // Create and push a test event
-  auto event = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "51", 0, "text"));
-  event->gtid = "uuid:51";
-
-  // Push event to queue
-  reader_->PushEvent(std::move(event));
-  EXPECT_EQ(reader_->GetQueueSize(), 1);
-
-  // Pop and process manually (simulating worker thread)
-  auto popped_event = reader_->PopEvent();
-  ASSERT_NE(popped_event, nullptr);
-
-  // Process the event
-  bool success = reader_->ProcessEvent(*popped_event);
-
-  // The bug: GTID and processed_events are updated regardless of success
-  // After fix: They should only be updated if success is true
-  if (success) {
-    // OK to update GTID
-    reader_->processed_events_++;
-    reader_->SetCurrentGTID(popped_event->gtid);
-    EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:51");
-    EXPECT_EQ(reader_->GetProcessedEvents(), initial_processed + 1);
-  } else {
-    // Should NOT update GTID - this is where the bug manifests
-    // The fix ensures GTID stays at original value
-    EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:50") << "Bug #3: GTID should not be updated on ProcessEvent failure";
-    EXPECT_EQ(reader_->GetProcessedEvents(), initial_processed)
-        << "Bug #3: processed_events should not be incremented on failure";
-  }
-}
-
-// =============================================================================
-// Bug #7: Graceful shutdown event loss
-// =============================================================================
-// The bug is in binlog_reader.cpp WorkerThreadFunc
-// while (!should_stop_) { ... }
-//
-// When should_stop_ becomes true, the loop exits immediately even if
-// there are pending events in the queue. This causes event loss during
-// graceful shutdown.
-//
-// The fix: Continue processing until queue is empty
-// =============================================================================
-
-/**
- * @brief Test that pending events are processed during shutdown
- *
- * This test verifies that when Stop() is called, the worker thread
- * processes all remaining events in the queue before exiting.
- */
-TEST_F(BinlogReaderBugFixTest, Bug7_PendingEventsProcessedDuringShutdown) {
-  // The buggy code in WorkerThreadFunc:
-  //   while (!should_stop_) {
-  //     auto event = PopEvent();
-  //     ...
-  //   }
-  //
-  // When should_stop_ becomes true (via Stop()), the loop exits immediately
-  // without processing any remaining events in the queue.
-  //
-  // The correct code should be:
-  //   while (true) {
-  //     auto event = PopEvent();
-  //     if (!event) break;  // PopEvent returns nullptr only when should_stop && queue.empty
-  //     ...
-  //   }
-
-  // Test the PopEvent behavior during shutdown
-  // 1. Push some events to the queue
-  auto event1 = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "1", 1, "text1"));
-  auto event2 = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "2", 1, "text2"));
-  auto event3 = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "3", 1, "text3"));
-
-  reader_->PushEvent(std::move(event1));
-  reader_->PushEvent(std::move(event2));
-  reader_->PushEvent(std::move(event3));
-
-  EXPECT_EQ(reader_->GetQueueSize(), 3);
-
-  // 2. Set should_stop_ flag (simulating Stop() call)
-  reader_->should_stop_ = true;
-
-  // 3. Verify PopEvent still returns events when queue is not empty
-  auto popped1 = reader_->PopEvent();
-  ASSERT_NE(popped1, nullptr) << "Bug #7: PopEvent should return event when queue is not empty, even during shutdown";
-  EXPECT_EQ(popped1->primary_key, "1");
-
-  auto popped2 = reader_->PopEvent();
-  ASSERT_NE(popped2, nullptr) << "Bug #7: PopEvent should return second event";
-  EXPECT_EQ(popped2->primary_key, "2");
-
-  auto popped3 = reader_->PopEvent();
-  ASSERT_NE(popped3, nullptr) << "Bug #7: PopEvent should return third event";
-  EXPECT_EQ(popped3->primary_key, "3");
-
-  // 4. Now queue is empty, PopEvent should return nullptr
-  auto popped4 = reader_->PopEvent();
-  EXPECT_EQ(popped4, nullptr) << "PopEvent should return nullptr when queue is empty and should_stop is true";
-}
-
-/**
- * @brief Test that the worker thread loop condition is correct
- *
- * The buggy loop:
- *   while (!should_stop_) { ... }
- *
- * The fixed loop:
- *   while (true) {
- *     auto event = PopEvent();
- *     if (!event) break;  // Only exits when both should_stop AND queue empty
- *     ...
- *   }
- */
-TEST_F(BinlogReaderBugFixTest, Bug7_WorkerLoopConditionTest) {
-  // Simulate worker thread behavior
-
-  // Test case 1: should_stop=false, queue has events -> continue processing
-  {
-    reader_->should_stop_ = false;
-    auto event = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "1", 1, "text"));
-    reader_->PushEvent(std::move(event));
-
-    auto popped = reader_->PopEvent();
-    ASSERT_NE(popped, nullptr) << "Should process events when not stopping";
-  }
-
-  // Test case 2: should_stop=true, queue has events -> should STILL process
-  {
-    // Note: PushEvent will not add events when should_stop_ is true (by design)
-    // So we need to add event while should_stop_ is false, then set should_stop_
-    reader_->should_stop_ = false;
-    auto event = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "2", 1, "text"));
-    reader_->PushEvent(std::move(event));
-    ASSERT_EQ(reader_->GetQueueSize(), 1) << "Event should be in queue";
-
-    // Now set should_stop_ and verify PopEvent still returns the event
-    reader_->should_stop_ = true;
-
-    auto popped = reader_->PopEvent();
-    // After fix: This should NOT be nullptr because queue is not empty
-    ASSERT_NE(popped, nullptr) << "Bug #7: Should process remaining events during shutdown";
-  }
-
-  // Test case 3: should_stop=true, queue empty -> exit
-  {
-    // Queue is already empty from previous pop
-    EXPECT_EQ(reader_->GetQueueSize(), 0);
-    auto popped = reader_->PopEvent();
-    EXPECT_EQ(popped, nullptr) << "Should exit when stopping and queue is empty";
-  }
-}
-
-// =============================================================================
-// Bug #8: Use-after-free during cleanup when Stop() is called multiple times
-// =============================================================================
-// The bug is in binlog_reader.cpp:280-293
-// If Stop() is called from multiple threads concurrently, there is a race
-// condition where threads can be joined/reset multiple times.
-//
-// The fix: Use mutex to protect Stop() or use compare_exchange to ensure
-// Stop() logic runs only once.
-// =============================================================================
-
-/**
- * @brief Test that multiple Stop() calls are handled safely
- *
- * Bug #8: Calling Stop() multiple times should be safe and not cause
- * use-after-free or double-join issues.
- */
-TEST_F(BinlogReaderBugFixTest, Bug8_MultipleStopCallsSafe) {
-  // Calling Stop() multiple times should be safe
-  reader_->Stop();
-  reader_->Stop();  // Second call should be a no-op
-  reader_->Stop();  // Third call should also be a no-op
-
-  // Should not crash and IsRunning() should be false
-  EXPECT_FALSE(reader_->IsRunning());
-}
-
-/**
- * @brief Test concurrent Stop() calls
- *
- * Bug #8: When multiple threads call Stop() concurrently, there should be
- * no race condition causing crashes or undefined behavior.
- */
-TEST_F(BinlogReaderBugFixTest, Bug8_ConcurrentStopCalls) {
-  // Launch multiple threads that all call Stop()
-  std::vector<std::thread> threads;
-  const int num_threads = 10;
-
-  for (int i = 0; i < num_threads; ++i) {
-    threads.emplace_back([this]() { reader_->Stop(); });
-  }
-
-  // Wait for all threads to complete
-  for (auto& t : threads) {
-    t.join();
-  }
-
-  // Should not crash and reader should be stopped
-  EXPECT_FALSE(reader_->IsRunning());
-}
-
-// =============================================================================
-// Bug #12: No schema change detection in TableMap cache
-// =============================================================================
-// The bug is that when a TABLE_MAP event arrives for a table already in cache,
-// the code doesn't detect if the schema changed. This is important for:
-// 1. Logging schema changes for debugging
-// 2. Ensuring the correct metadata is used for row parsing
-//
-// The fix: Use AddOrUpdate method that detects schema changes and logs them.
-// =============================================================================
-
-/**
- * @brief Test fixture for TableMetadataCache tests
- */
-class TableMetadataCacheTest : public ::testing::Test {
- protected:
-  TableMetadataCache cache_;
-
-  TableMetadata CreateMetadata(uint64_t table_id, const std::string& db, const std::string& table,
-                               std::vector<std::pair<std::string, ColumnType>> columns) {
-    TableMetadata meta;
-    meta.table_id = table_id;
-    meta.database_name = db;
-    meta.table_name = table;
-    for (const auto& [name, type] : columns) {
-      ColumnMetadata col;
-      col.name = name;
-      col.type = type;
-      col.metadata = 0;
-      meta.columns.push_back(col);
-    }
-    return meta;
-  }
-};
-
-/**
- * @test Bug #12: AddOrUpdate should detect new entries
- */
-TEST_F(TableMetadataCacheTest, Bug12_AddOrUpdateDetectsNewEntry) {
-  auto meta = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  auto result = cache_.AddOrUpdate(100, meta);
-  EXPECT_EQ(TableMetadataCache::AddResult::kAdded, result);
-
-  // Verify the entry was added
-  const auto* cached = cache_.Get(100);
-  ASSERT_NE(nullptr, cached);
-  EXPECT_EQ("users", cached->table_name);
-}
-
-/**
- * @test Bug #12: AddOrUpdate should detect schema unchanged
- */
-TEST_F(TableMetadataCacheTest, Bug12_AddOrUpdateDetectsSameSchema) {
-  auto meta = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  cache_.AddOrUpdate(100, meta);
-
-  // Same schema should return kUpdated
-  auto result = cache_.AddOrUpdate(100, meta);
-  EXPECT_EQ(TableMetadataCache::AddResult::kUpdated, result);
-}
-
-/**
- * @test Bug #12: AddOrUpdate should detect column count change
- */
-TEST_F(TableMetadataCacheTest, Bug12_AddOrUpdateDetectsColumnCountChange) {
-  auto meta1 = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  cache_.AddOrUpdate(100, meta1);
-
-  // Add a new column (ALTER TABLE ADD COLUMN)
-  auto meta2 =
-      CreateMetadata(100, "test_db", "users",
-                     {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}, {"email", ColumnType::VARCHAR}});
-
-  auto result = cache_.AddOrUpdate(100, meta2);
-  EXPECT_EQ(TableMetadataCache::AddResult::kSchemaChanged, result);
-
-  // Verify the cache was updated
-  const auto* cached = cache_.Get(100);
-  ASSERT_NE(nullptr, cached);
-  EXPECT_EQ(3, cached->columns.size());
-}
-
-/**
- * @test Bug #12: AddOrUpdate should detect column type change
- */
-TEST_F(TableMetadataCacheTest, Bug12_AddOrUpdateDetectsColumnTypeChange) {
-  auto meta1 = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  cache_.AddOrUpdate(100, meta1);
-
-  // Change column type (ALTER TABLE MODIFY COLUMN)
-  auto meta2 = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONGLONG}, {"name", ColumnType::VARCHAR}});
-
-  auto result = cache_.AddOrUpdate(100, meta2);
-  EXPECT_EQ(TableMetadataCache::AddResult::kSchemaChanged, result);
-}
-
-/**
- * @test Bug #12: AddOrUpdate should detect column name change
- */
-TEST_F(TableMetadataCacheTest, Bug12_AddOrUpdateDetectsColumnNameChange) {
-  auto meta1 = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  cache_.AddOrUpdate(100, meta1);
-
-  // Rename column (ALTER TABLE CHANGE COLUMN)
-  auto meta2 = CreateMetadata(100, "test_db", "users", {{"user_id", ColumnType::LONG}, {"name", ColumnType::VARCHAR}});
-
-  auto result = cache_.AddOrUpdate(100, meta2);
-  EXPECT_EQ(TableMetadataCache::AddResult::kSchemaChanged, result);
-}
-
-/**
- * @test Bug #12: Contains method should work correctly
- */
-TEST_F(TableMetadataCacheTest, Bug12_ContainsMethod) {
-  EXPECT_FALSE(cache_.Contains(100));
-
-  auto meta = CreateMetadata(100, "test_db", "users", {{"id", ColumnType::LONG}});
-  cache_.AddOrUpdate(100, meta);
-
-  EXPECT_TRUE(cache_.Contains(100));
-  EXPECT_FALSE(cache_.Contains(101));
-}
-
-#endif  // USE_MYSQL
