--- tests/mysql/binlog_reader_test.cpp
@@ -1,1410 +1,0 @@
-/**
- * @file binlog_reader_test.cpp
- * @brief Unit tests for binlog reader
- */
-
-#include <gtest/gtest.h>
-
-#include <atomic>
-#include <chrono>
-#include <thread>
-
-#define private public
-#define protected public
-#include "mysql/binlog_reader.h"
-#undef private
-#undef protected
-
-#ifdef USE_MYSQL
-
-#include "mock_connection.h"
-#include "mysql/binlog_filter_evaluator.h"
-#include "server/server_stats.h"
-#include "server/server_types.h"
-
-using namespace mygramdb::mysql;
-using namespace mygramdb;
-
-namespace {
-
-/**
- * @brief Helper that creates a default table configuration for tests
- */
-config::TableConfig MakeDefaultTableConfig() {
-  config::TableConfig table_config;
-  table_config.name = "articles";
-  table_config.primary_key = "id";
-  table_config.text_source.column = "content";
-
-  config::RequiredFilterConfig required_filter;
-  required_filter.name = "status";
-  required_filter.type = "int";
-  required_filter.op = "=";
-  required_filter.value = "1";
-  table_config.required_filters.push_back(required_filter);
-
-  config::FilterConfig optional_filter;
-  optional_filter.name = "category";
-  optional_filter.type = "string";
-  table_config.filters.push_back(optional_filter);
-
-  return table_config;
-}
-
-/**
- * @brief BinlogReader test fixture providing in-memory dependencies
- */
-class BinlogReaderFixture : public ::testing::Test {
- protected:
-  BinlogReaderFixture() : connection_(connection_config_), index_(2) {}
-
-  void SetUp() override {
-    table_config_ = MakeDefaultTableConfig();
-    reader_config_.start_gtid = "uuid:1";
-    reader_config_.queue_size = 2;
-    reader_config_.reconnect_delay_ms = 10;
-    reader_config_.server_id = 12345;  // Test server ID
-
-    index_.Clear();
-    doc_store_.Clear();
-    ResetReader();
-  }
-
-  void TearDown() override {
-    reader_.reset();
-    doc_store_.Clear();
-    index_.Clear();
-  }
-
-  /**
-   * @brief Recreate BinlogReader with current configuration
-   */
-  void ResetReader() {
-    reader_ = std::make_unique<BinlogReader>(connection_, index_, doc_store_, table_config_, reader_config_);
-  }
-
-  /**
-   * @brief Utility to build a fully populated event for tests
-   */
-  BinlogEvent MakeEvent(BinlogEventType type, const std::string& pk, int status, const std::string& text = "hello") {
-    BinlogEvent event;
-    event.type = type;
-    event.table_name = table_config_.name;
-    event.primary_key = pk;
-    event.text = text;
-    event.gtid = "uuid:" + pk;
-    event.filters["status"] = static_cast<int64_t>(status);
-    event.filters["category"] = std::string("news");
-    return event;
-  }
-
-  Connection::Config connection_config_;
-  Connection connection_;
-  index::Index index_;
-  storage::DocumentStore doc_store_;
-  config::TableConfig table_config_;
-  BinlogReader::Config reader_config_;
-  std::unique_ptr<BinlogReader> reader_;
-};
-
-}  // namespace
-
-/**
- * @brief Validate Start/Stop lifecycle without a real MySQL connection
- */
-TEST_F(BinlogReaderFixture, StartStopLifecycleWithoutConnection) {
-  EXPECT_FALSE(reader_->IsRunning());
-  EXPECT_FALSE(reader_->Start());
-  EXPECT_FALSE(reader_->IsRunning());
-  EXPECT_NE(reader_->GetLastError().find("connection not established"), std::string::npos);
-
-  reader_->Stop();
-  EXPECT_FALSE(reader_->IsRunning());
-
-  // Calling Stop multiple times should be safe
-  reader_->Stop();
-  EXPECT_FALSE(reader_->IsRunning());
-}
-
-/**
- * @brief Ensure Start reports an error when the reader is already running
- */
-TEST_F(BinlogReaderFixture, RejectsDoubleStart) {
-  reader_->running_ = true;
-  EXPECT_FALSE(reader_->Start());
-  EXPECT_NE(reader_->GetLastError().find("already running"), std::string::npos);
-}
-
-/**
- * @brief Exercise queue push/pop helpers without worker threads
- */
-TEST_F(BinlogReaderFixture, PushAndPopEvents) {
-  auto first = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "1", 1));
-  reader_->PushEvent(std::move(first));
-  EXPECT_EQ(reader_->GetQueueSize(), 1);
-
-  auto popped = reader_->PopEvent();
-  ASSERT_NE(popped, nullptr);
-  EXPECT_EQ(popped->primary_key, "1");
-  EXPECT_EQ(reader_->GetQueueSize(), 0);
-}
-
-/**
- * @brief Verify PushEvent blocks when queue is full until space becomes available
- */
-TEST_F(BinlogReaderFixture, PushBlocksWhenQueueFull) {
-  reader_->config_.queue_size = 1;
-  auto first = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "1", 1));
-  reader_->PushEvent(std::move(first));
-
-  std::atomic<bool> second_pushed{false};
-  std::thread producer([&] {
-    auto second = std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "2", 1));
-    reader_->PushEvent(std::move(second));
-    second_pushed.store(true);
-  });
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(20));
-  EXPECT_FALSE(second_pushed.load());
-
-  auto popped = reader_->PopEvent();
-  ASSERT_NE(popped, nullptr);
-  producer.join();
-  EXPECT_TRUE(second_pushed.load());
-
-  // Drain queue for subsequent tests
-  popped = reader_->PopEvent();
-  ASSERT_NE(popped, nullptr);
-  EXPECT_EQ(reader_->GetQueueSize(), 0);
-}
-
-/**
- * @brief Ensure PopEvent blocks until a producer pushes data
- */
-TEST_F(BinlogReaderFixture, PopBlocksUntilEventArrives) {
-  std::atomic<bool> pop_completed{false};
-  std::thread consumer([&] {
-    auto event = reader_->PopEvent();
-    pop_completed.store(event != nullptr);
-    if (event) {
-      EXPECT_EQ(event->primary_key, "7");
-    }
-  });
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(20));
-  EXPECT_FALSE(pop_completed.load());
-
-  reader_->PushEvent(std::make_unique<BinlogEvent>(MakeEvent(BinlogEventType::INSERT, "7", 1)));
-  consumer.join();
-  EXPECT_TRUE(pop_completed.load());
-  EXPECT_EQ(reader_->GetQueueSize(), 0);
-}
-
-/**
- * @brief Confirm PopEvent unblocks and returns nullptr when reader is stopped
- */
-TEST_F(BinlogReaderFixture, PopReturnsFalseWhenStopping) {
-  std::atomic<bool> pop_result{true};
-  std::thread consumer([&] {
-    auto event = reader_->PopEvent();
-    pop_result.store(event == nullptr);
-  });
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(20));
-  reader_->should_stop_ = true;
-  reader_->queue_cv_.notify_all();
-
-  consumer.join();
-  EXPECT_TRUE(pop_result.load());  // Should be nullptr (true) when stopping
-  reader_->should_stop_ = false;
-}
-
-/**
- * @brief Validate INSERT events create documents when filters match
- */
-TEST_F(BinlogReaderFixture, ProcessInsertAddsDocument) {
-  BinlogEvent insert_event = MakeEvent(BinlogEventType::INSERT, "42", 1, "Breaking news");
-  ASSERT_TRUE(reader_->ProcessEvent(insert_event));
-
-  auto doc_id = doc_store_.GetDocId("42");
-  ASSERT_TRUE(doc_id.has_value());
-  auto stored_doc = doc_store_.GetDocument(doc_id.value());
-  ASSERT_TRUE(stored_doc.has_value());
-  EXPECT_EQ(std::get<std::string>(stored_doc->filters["category"]), "news");
-}
-
-/**
- * @brief Ensure UPDATE removes rows when they no longer satisfy required filters
- */
-TEST_F(BinlogReaderFixture, ProcessUpdateRemovesWhenFiltersFail) {
-  ASSERT_TRUE(reader_->ProcessEvent(MakeEvent(BinlogEventType::INSERT, "90", 1, "Initial")));
-
-  BinlogEvent update_event = MakeEvent(BinlogEventType::UPDATE, "90", 0, "Updated text");
-  ASSERT_TRUE(reader_->ProcessEvent(update_event));
-  EXPECT_FALSE(doc_store_.GetDocId("90").has_value());
-}
-
-/**
- * @brief Test UPDATE properly updates full-text index when text changes
- *
- * Verifies that when an UPDATE event changes the text content:
- * 1. The old text is removed from the index using old_text field
- * 2. The new text is added to the index
- * 3. Document store filters are updated
- */
-TEST_F(BinlogReaderFixture, ProcessUpdateUpdatesIndexWithTextChange) {
-  // Insert initial document with text "hello world"
-  ASSERT_TRUE(reader_->ProcessEvent(MakeEvent(BinlogEventType::INSERT, "100", 1, "hello world")));
-
-  auto doc_id = doc_store_.GetDocId("100");
-  ASSERT_TRUE(doc_id.has_value());
-
-  // Verify initial text is in the index (bigram "he" from "hello")
-  EXPECT_GT(index_.Count("he"), 0);
-
-  // Create UPDATE event with new text "goodbye universe"
-  BinlogEvent update_event = MakeEvent(BinlogEventType::UPDATE, "100", 1, "goodbye universe");
-  update_event.old_text = "hello world";  // Set old_text for index update
-
-  ASSERT_TRUE(reader_->ProcessEvent(update_event));
-
-  // Verify document still exists (not removed and re-added)
-  auto updated_doc_id = doc_store_.GetDocId("100");
-  ASSERT_TRUE(updated_doc_id.has_value());
-  EXPECT_EQ(updated_doc_id.value(), doc_id.value());
-
-  // Verify old text was removed from index (bigram "he" from "hello" should be gone)
-  EXPECT_EQ(index_.Count("he"), 0);
-
-  // Verify new text was added to index (bigram "go" from "goodbye" should exist)
-  EXPECT_GT(index_.Count("go"), 0);
-
-  // Verify filters were updated
-  auto stored_doc = doc_store_.GetDocument(doc_id.value());
-  ASSERT_TRUE(stored_doc.has_value());
-  EXPECT_EQ(std::get<std::string>(stored_doc->filters["category"]), "news");
-  EXPECT_EQ(std::get<int64_t>(stored_doc->filters["status"]), 1);
-}
-
-/**
- * @brief Test UPDATE handles empty old_text gracefully
- *
- * Ensures that if old_text is empty (shouldn't happen in practice with proper
- * before image parsing, but defensive), the update still works and adds new text.
- */
-TEST_F(BinlogReaderFixture, ProcessUpdateHandlesEmptyOldText) {
-  // Insert initial document
-  ASSERT_TRUE(reader_->ProcessEvent(MakeEvent(BinlogEventType::INSERT, "101", 1, "original text")));
-
-  auto doc_id = doc_store_.GetDocId("101");
-  ASSERT_TRUE(doc_id.has_value());
-
-  // Create UPDATE event with empty old_text
-  BinlogEvent update_event = MakeEvent(BinlogEventType::UPDATE, "101", 1, "newtext");
-  update_event.old_text = "";  // Empty old_text
-
-  // Should still process successfully
-  ASSERT_TRUE(reader_->ProcessEvent(update_event));
-
-  // Verify document still exists
-  auto updated_doc_id = doc_store_.GetDocId("101");
-  ASSERT_TRUE(updated_doc_id.has_value());
-  EXPECT_EQ(updated_doc_id.value(), doc_id.value());
-
-  // Verify new text was added to index (bigram "ne" from "newtext")
-  EXPECT_GT(index_.Count("ne"), 0);
-
-  // Verify filters are preserved
-  auto stored_doc = doc_store_.GetDocument(doc_id.value());
-  ASSERT_TRUE(stored_doc.has_value());
-  EXPECT_EQ(std::get<std::string>(stored_doc->filters["category"]), "news");
-}
-
-/**
- * @brief Test UPDATE when only filters change (no text change)
- *
- * Verifies that UPDATE correctly handles cases where only filter values change
- * but the text content remains the same. Index should update (remove old, add same)
- * but content remains searchable.
- */
-TEST_F(BinlogReaderFixture, ProcessUpdateOnlyFiltersChange) {
-  // Insert initial document
-  BinlogEvent insert_event = MakeEvent(BinlogEventType::INSERT, "102", 1, "sametext");
-  insert_event.filters["category"] = std::string("sports");
-  ASSERT_TRUE(reader_->ProcessEvent(insert_event));
-
-  auto doc_id = doc_store_.GetDocId("102");
-  ASSERT_TRUE(doc_id.has_value());
-
-  // Verify initial category and text is indexed (bigram "sa" from "sametext")
-  auto initial_doc = doc_store_.GetDocument(doc_id.value());
-  ASSERT_TRUE(initial_doc.has_value());
-  EXPECT_EQ(std::get<std::string>(initial_doc->filters["category"]), "sports");
-  EXPECT_GT(index_.Count("sa"), 0);
-
-  // Update with same text but different filter
-  BinlogEvent update_event = MakeEvent(BinlogEventType::UPDATE, "102", 1, "sametext");
-  update_event.old_text = "sametext";                      // Same text
-  update_event.filters["category"] = std::string("news");  // Different category
-
-  ASSERT_TRUE(reader_->ProcessEvent(update_event));
-
-  // Verify document still exists (same doc_id)
-  auto updated_doc_id = doc_store_.GetDocId("102");
-  ASSERT_TRUE(updated_doc_id.has_value());
-  EXPECT_EQ(updated_doc_id.value(), doc_id.value());
-
-  // Verify text is still in index (same text was removed and re-added)
-  EXPECT_GT(index_.Count("sa"), 0);
-
-  // Verify filters were updated
-  auto stored_doc = doc_store_.GetDocument(doc_id.value());
-  ASSERT_TRUE(stored_doc.has_value());
-  EXPECT_EQ(std::get<std::string>(stored_doc->filters["category"]), "news");
-}
-
-/**
- * @brief Verify DELETE events remove documents and index entries
- */
-TEST_F(BinlogReaderFixture, ProcessDeleteRemovesDocument) {
-  ASSERT_TRUE(reader_->ProcessEvent(MakeEvent(BinlogEventType::INSERT, "77", 1, "Row")));
-
-  BinlogEvent delete_event = MakeEvent(BinlogEventType::DELETE, "77", 1, "Row");
-  ASSERT_TRUE(reader_->ProcessEvent(delete_event));
-  EXPECT_FALSE(doc_store_.GetDocId("77").has_value());
-}
-
-/**
- * @brief Validate DDL TRUNCATE clears index and store
- */
-TEST_F(BinlogReaderFixture, ProcessDdlTruncateClearsState) {
-  ASSERT_TRUE(reader_->ProcessEvent(MakeEvent(BinlogEventType::INSERT, "5", 1, "Body")));
-  EXPECT_EQ(doc_store_.Size(), 1);
-
-  BinlogEvent ddl_event;
-  ddl_event.type = BinlogEventType::DDL;
-  ddl_event.table_name = table_config_.name;
-  ddl_event.text = "TRUNCATE TABLE articles";
-  ASSERT_TRUE(reader_->ProcessEvent(ddl_event));
-  EXPECT_EQ(doc_store_.Size(), 0);
-}
-
-/**
- * @brief Confirm events missing required filters are skipped
- */
-TEST_F(BinlogReaderFixture, SkipsEventsMissingRequiredFilters) {
-  BinlogEvent insert_event = MakeEvent(BinlogEventType::INSERT, "21", 1, "Text");
-  insert_event.filters.erase("status");
-  ASSERT_TRUE(reader_->ProcessEvent(insert_event));
-  EXPECT_FALSE(doc_store_.GetDocId("21").has_value());
-}
-
-/**
- * @brief Exercise GTID setters/getters
- */
-TEST_F(BinlogReaderFixture, TracksGtidUpdates) {
-  reader_->SetCurrentGTID("uuid:10");
-  EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:10");
-  reader_->UpdateCurrentGTID("uuid:11");
-  EXPECT_EQ(reader_->GetCurrentGTID(), "uuid:11");
-}
-
-/**
- * @brief Ensure events are routed to correct TableContext in multi-table mode
- */
-TEST_F(BinlogReaderFixture, MultiTableProcessesCorrectTable) {
-  server::TableContext articles_ctx;
-  articles_ctx.name = "articles";
-  articles_ctx.config = table_config_;
-  articles_ctx.index = std::make_unique<index::Index>();
-  articles_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  server::TableContext comments_ctx;
-  comments_ctx.name = "comments";
-  comments_ctx.config = table_config_;
-  comments_ctx.config.name = "comments";
-  comments_ctx.index = std::make_unique<index::Index>();
-  comments_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  std::unordered_map<std::string, server::TableContext*> contexts = {
-      {articles_ctx.name, &articles_ctx},
-      {comments_ctx.name, &comments_ctx},
-  };
-
-  BinlogReader multi_reader(connection_, contexts, reader_config_);
-
-  BinlogEvent comment_event = MakeEvent(BinlogEventType::INSERT, "300", 1, "Comment");
-  comment_event.table_name = "comments";
-  ASSERT_TRUE(multi_reader.ProcessEvent(comment_event));
-  EXPECT_TRUE(comments_ctx.doc_store->GetDocId("300").has_value());
-  EXPECT_FALSE(articles_ctx.doc_store->GetDocId("300").has_value());
-}
-
-/**
- * @brief Multi-table mode should ignore tables that are not tracked
- */
-TEST_F(BinlogReaderFixture, MultiTableSkipsUnknownTable) {
-  server::TableContext articles_ctx;
-  articles_ctx.name = "articles";
-  articles_ctx.config = table_config_;
-  articles_ctx.index = std::make_unique<index::Index>();
-  articles_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  std::unordered_map<std::string, server::TableContext*> contexts = {
-      {articles_ctx.name, &articles_ctx},
-  };
-
-  BinlogReader multi_reader(connection_, contexts, reader_config_);
-  BinlogEvent other_event = MakeEvent(BinlogEventType::INSERT, "400", 1, "Ignored");
-  other_event.table_name = "not_tracked";
-  ASSERT_TRUE(multi_reader.ProcessEvent(other_event));
-  EXPECT_EQ(articles_ctx.doc_store->Size(), 0);
-}
-
-/**
- * @brief Test BinlogEvent structure
- */
-TEST(BinlogReaderTest, EventStructure) {
-  BinlogEvent event;
-  event.type = BinlogEventType::INSERT;
-  event.table_name = "test_table";
-  event.primary_key = "123";
-  event.text = "test text";
-  event.gtid = "uuid:1";
-
-  EXPECT_EQ(event.type, BinlogEventType::INSERT);
-  EXPECT_EQ(event.table_name, "test_table");
-  EXPECT_EQ(event.primary_key, "123");
-  EXPECT_EQ(event.text, "test text");
-  EXPECT_EQ(event.gtid, "uuid:1");
-}
-
-/**
- * @brief Test BinlogEventType enum
- */
-TEST(BinlogReaderTest, EventTypes) {
-  EXPECT_NE(BinlogEventType::INSERT, BinlogEventType::UPDATE);
-  EXPECT_NE(BinlogEventType::INSERT, BinlogEventType::DELETE);
-  EXPECT_NE(BinlogEventType::UPDATE, BinlogEventType::DELETE);
-  EXPECT_NE(BinlogEventType::INSERT, BinlogEventType::UNKNOWN);
-}
-
-/**
- * @brief Test BinlogReader construction
- */
-TEST(BinlogReaderTest, Construction) {
-  // Create dependencies
-  Connection::Config conn_config;
-  conn_config.host = "localhost";
-  conn_config.user = "test";
-  conn_config.password = "test";
-
-  Connection conn(conn_config);
-
-  index::Index idx(1);
-  storage::DocumentStore doc_store;
-
-  config::TableConfig table_config;
-  table_config.name = "test_table";
-  table_config.primary_key = "id";
-
-  BinlogReader::Config reader_config;
-  reader_config.start_gtid = "uuid:1";
-  reader_config.queue_size = 1000;
-  reader_config.server_id = 12345;  // Test server ID
-
-  BinlogReader reader(conn, idx, doc_store, table_config, reader_config);
-
-  // Should construct successfully
-  EXPECT_FALSE(reader.IsRunning());
-  EXPECT_EQ(reader.GetProcessedEvents(), 0);
-  EXPECT_EQ(reader.GetQueueSize(), 0);
-}
-
-/**
- * @brief Test BinlogReader initial state
- */
-TEST(BinlogReaderTest, InitialState) {
-  Connection::Config conn_config;
-  Connection conn(conn_config);
-
-  index::Index idx(1);
-  storage::DocumentStore doc_store;
-
-  config::TableConfig table_config;
-  table_config.name = "test_table";
-
-  BinlogReader::Config reader_config;
-  reader_config.start_gtid = "3E11FA47-71CA-11E1-9E33-C80AA9429562:100";
-  reader_config.server_id = 12345;  // Test server ID
-
-  BinlogReader reader(conn, idx, doc_store, table_config, reader_config);
-
-  EXPECT_FALSE(reader.IsRunning());
-  EXPECT_EQ(reader.GetCurrentGTID(), "3E11FA47-71CA-11E1-9E33-C80AA9429562:100");
-  EXPECT_EQ(reader.GetQueueSize(), 0);
-  EXPECT_EQ(reader.GetProcessedEvents(), 0);
-}
-
-/**
- * @brief Test BinlogReader config
- */
-TEST(BinlogReaderTest, Config) {
-  BinlogReader::Config config;
-
-  // Default values
-  EXPECT_EQ(config.queue_size, 10000);
-  EXPECT_EQ(config.reconnect_delay_ms, 1000);
-  EXPECT_EQ(config.server_id, 0);  // Default server_id is 0 (invalid for replication)
-
-  // Custom values
-  config.start_gtid = "test:123";
-  config.queue_size = 5000;
-  config.reconnect_delay_ms = 500;
-  config.server_id = 12345;
-
-  EXPECT_EQ(config.start_gtid, "test:123");
-  EXPECT_EQ(config.queue_size, 5000);
-  EXPECT_EQ(config.reconnect_delay_ms, 500);
-  EXPECT_EQ(config.server_id, 12345);
-}
-
-/**
- * @brief Test BinlogEvent with filters
- */
-TEST(BinlogReaderTest, EventWithFilters) {
-  BinlogEvent event;
-  event.type = BinlogEventType::INSERT;
-  event.table_name = "articles";
-  event.primary_key = "456";
-  event.text = "article text";
-
-  // Add filters
-  event.filters["status"] = static_cast<int64_t>(1);
-  event.filters["category"] = std::string("news");
-
-  EXPECT_EQ(event.filters.size(), 2);
-
-  auto status = std::get<int64_t>(event.filters["status"]);
-  auto category = std::get<std::string>(event.filters["category"]);
-
-  EXPECT_EQ(status, 1);
-  EXPECT_EQ(category, "news");
-}
-
-/**
- * @brief Test multiple event types
- */
-TEST(BinlogReaderTest, MultipleEventTypes) {
-  BinlogEvent insert_event;
-  insert_event.type = BinlogEventType::INSERT;
-  insert_event.primary_key = "1";
-
-  BinlogEvent update_event;
-  update_event.type = BinlogEventType::UPDATE;
-  update_event.primary_key = "2";
-
-  BinlogEvent delete_event;
-  delete_event.type = BinlogEventType::DELETE;
-  delete_event.primary_key = "3";
-
-  EXPECT_EQ(insert_event.type, BinlogEventType::INSERT);
-  EXPECT_EQ(update_event.type, BinlogEventType::UPDATE);
-  EXPECT_EQ(delete_event.type, BinlogEventType::DELETE);
-
-  EXPECT_NE(insert_event.primary_key, update_event.primary_key);
-  EXPECT_NE(update_event.primary_key, delete_event.primary_key);
-}
-
-/**
- * @brief Test DDL event type
- */
-TEST(BinlogReaderTest, DDLEventType) {
-  BinlogEvent ddl_event;
-  ddl_event.type = BinlogEventType::DDL;
-  ddl_event.table_name = "test_table";
-  ddl_event.text = "TRUNCATE TABLE test_table";
-
-  EXPECT_EQ(ddl_event.type, BinlogEventType::DDL);
-  EXPECT_EQ(ddl_event.table_name, "test_table");
-  EXPECT_EQ(ddl_event.text, "TRUNCATE TABLE test_table");
-
-  // DDL events should be distinct from other event types
-  EXPECT_NE(BinlogEventType::DDL, BinlogEventType::INSERT);
-  EXPECT_NE(BinlogEventType::DDL, BinlogEventType::UPDATE);
-  EXPECT_NE(BinlogEventType::DDL, BinlogEventType::DELETE);
-  EXPECT_NE(BinlogEventType::DDL, BinlogEventType::UNKNOWN);
-}
-
-/**
- * @brief Test TRUNCATE TABLE DDL event
- */
-TEST(BinlogReaderTest, TruncateTableEvent) {
-  BinlogEvent event;
-  event.type = BinlogEventType::DDL;
-  event.table_name = "articles";
-  event.text = "TRUNCATE TABLE articles";
-
-  EXPECT_EQ(event.type, BinlogEventType::DDL);
-  EXPECT_NE(event.text.find("TRUNCATE"), std::string::npos);
-}
-
-/**
- * @brief Test ALTER TABLE DDL event
- */
-TEST(BinlogReaderTest, AlterTableEvent) {
-  BinlogEvent event;
-  event.type = BinlogEventType::DDL;
-  event.table_name = "users";
-  event.text = "ALTER TABLE users ADD COLUMN email VARCHAR(255)";
-
-  EXPECT_EQ(event.type, BinlogEventType::DDL);
-  EXPECT_NE(event.text.find("ALTER"), std::string::npos);
-}
-
-/**
- * @brief Test DROP TABLE DDL event
- */
-TEST(BinlogReaderTest, DropTableEvent) {
-  BinlogEvent event;
-  event.type = BinlogEventType::DDL;
-  event.table_name = "temp_table";
-  event.text = "DROP TABLE temp_table";
-
-  EXPECT_EQ(event.type, BinlogEventType::DDL);
-  EXPECT_NE(event.text.find("DROP"), std::string::npos);
-}
-
-/**
- * @brief Test DDL event with GTID
- */
-TEST(BinlogReaderTest, DDLEventWithGTID) {
-  BinlogEvent event;
-  event.type = BinlogEventType::DDL;
-  event.table_name = "products";
-  event.text = "TRUNCATE TABLE products";
-  event.gtid = "3E11FA47-71CA-11E1-9E33-C80AA9429562:150";
-
-  EXPECT_EQ(event.type, BinlogEventType::DDL);
-  EXPECT_EQ(event.gtid, "3E11FA47-71CA-11E1-9E33-C80AA9429562:150");
-  EXPECT_FALSE(event.gtid.empty());
-}
-
-/**
- * @brief Test various DDL statement formats
- */
-TEST(BinlogReaderTest, VariousDDLFormats) {
-  // Test case variations
-  BinlogEvent truncate_upper;
-  truncate_upper.type = BinlogEventType::DDL;
-  truncate_upper.text = "TRUNCATE TABLE MY_TABLE";
-  EXPECT_NE(truncate_upper.text.find("TRUNCATE"), std::string::npos);
-
-  BinlogEvent truncate_lower;
-  truncate_lower.type = BinlogEventType::DDL;
-  truncate_lower.text = "truncate table my_table";
-  EXPECT_NE(truncate_lower.text.find("truncate"), std::string::npos);
-
-  BinlogEvent alter_add_column;
-  alter_add_column.type = BinlogEventType::DDL;
-  alter_add_column.text = "ALTER TABLE users ADD COLUMN status INT";
-  EXPECT_NE(alter_add_column.text.find("ALTER"), std::string::npos);
-
-  BinlogEvent alter_modify_column;
-  alter_modify_column.type = BinlogEventType::DDL;
-  alter_modify_column.text = "ALTER TABLE users MODIFY COLUMN name VARCHAR(100)";
-  EXPECT_NE(alter_modify_column.text.find("MODIFY"), std::string::npos);
-
-  BinlogEvent drop_if_exists;
-  drop_if_exists.type = BinlogEventType::DDL;
-  drop_if_exists.text = "DROP TABLE IF EXISTS temp_table";
-  EXPECT_NE(drop_if_exists.text.find("DROP"), std::string::npos);
-}
-
-/**
- * @brief Test DDL event distinguishing from DML events
- */
-TEST(BinlogReaderTest, DDLvsDMLEvents) {
-  BinlogEvent dml_insert;
-  dml_insert.type = BinlogEventType::INSERT;
-  dml_insert.primary_key = "100";
-  dml_insert.text = "new record text";
-
-  BinlogEvent ddl_truncate;
-  ddl_truncate.type = BinlogEventType::DDL;
-  ddl_truncate.text = "TRUNCATE TABLE test_table";
-
-  // DDL events don't have primary keys (they affect entire table)
-  EXPECT_FALSE(dml_insert.primary_key.empty());
-  EXPECT_TRUE(ddl_truncate.primary_key.empty());
-
-  // DDL events store SQL query in text field
-  EXPECT_EQ(dml_insert.type, BinlogEventType::INSERT);
-  EXPECT_EQ(ddl_truncate.type, BinlogEventType::DDL);
-}
-
-/**
- * @brief Test BinlogReader with ServerStats integration
- */
-TEST(BinlogReaderTest, ServerStatsIntegration) {
-  Connection::Config conn_config;
-  conn_config.host = "localhost";
-  conn_config.user = "test";
-  conn_config.password = "test";
-
-  Connection conn(conn_config);
-
-  index::Index idx(1);
-  storage::DocumentStore doc_store;
-
-  config::TableConfig table_config;
-  table_config.name = "test_table";
-  table_config.primary_key = "id";
-
-  BinlogReader::Config reader_config;
-  reader_config.start_gtid = "uuid:1";
-
-  server::ServerStats stats;
-
-  // Create BinlogReader with ServerStats
-  BinlogReader reader(conn, idx, doc_store, table_config, reader_config, &stats);
-
-  // Verify initial statistics are zero
-  EXPECT_EQ(stats.GetReplInsertsApplied(), 0);
-  EXPECT_EQ(stats.GetReplInsertsSkipped(), 0);
-  EXPECT_EQ(stats.GetReplUpdatesApplied(), 0);
-  EXPECT_EQ(stats.GetReplDeletesApplied(), 0);
-
-  // Verify BinlogReader is not running
-  EXPECT_FALSE(reader.IsRunning());
-}
-
-/**
- * @brief Test BinlogReader SetServerStats method
- */
-TEST(BinlogReaderTest, SetServerStats) {
-  Connection::Config conn_config;
-  Connection conn(conn_config);
-
-  index::Index idx(1);
-  storage::DocumentStore doc_store;
-
-  config::TableConfig table_config;
-  table_config.name = "test_table";
-
-  BinlogReader::Config reader_config;
-
-  // Create BinlogReader without ServerStats
-  BinlogReader reader(conn, idx, doc_store, table_config, reader_config);
-
-  // Create ServerStats and set it
-  server::ServerStats stats;
-  reader.SetServerStats(&stats);
-
-  // Verify initial statistics are zero
-  EXPECT_EQ(stats.GetReplInsertsApplied(), 0);
-  EXPECT_EQ(stats.GetReplUpdatesApplied(), 0);
-  EXPECT_EQ(stats.GetReplDeletesApplied(), 0);
-}
-
-/**
- * @brief Test BinlogReader multi-table mode with ServerStats
- */
-TEST(BinlogReaderTest, MultiTableModeWithServerStats) {
-  Connection::Config conn_config;
-  Connection conn(conn_config);
-
-  // Create table contexts
-  server::TableContext table_ctx1;
-  table_ctx1.name = "table1";
-  table_ctx1.config.name = "table1";
-  table_ctx1.config.primary_key = "id";
-  table_ctx1.index = std::make_unique<index::Index>(1);
-  table_ctx1.doc_store = std::make_unique<storage::DocumentStore>();
-
-  server::TableContext table_ctx2;
-  table_ctx2.name = "table2";
-  table_ctx2.config.name = "table2";
-  table_ctx2.config.primary_key = "id";
-  table_ctx2.index = std::make_unique<index::Index>(1);
-  table_ctx2.doc_store = std::make_unique<storage::DocumentStore>();
-
-  std::unordered_map<std::string, server::TableContext*> table_contexts;
-  table_contexts["table1"] = &table_ctx1;
-  table_contexts["table2"] = &table_ctx2;
-
-  BinlogReader::Config reader_config;
-  server::ServerStats stats;
-
-  // Create BinlogReader in multi-table mode with ServerStats
-  BinlogReader reader(conn, table_contexts, reader_config, &stats);
-
-  // Verify initial statistics are zero
-  EXPECT_EQ(stats.GetReplInsertsApplied(), 0);
-  EXPECT_EQ(stats.GetReplUpdatesApplied(), 0);
-  EXPECT_EQ(stats.GetReplDeletesApplied(), 0);
-  EXPECT_EQ(stats.GetReplEventsSkippedOtherTables(), 0);
-
-  // Verify BinlogReader is not running
-  EXPECT_FALSE(reader.IsRunning());
-}
-
-/**
- * @brief Test BinlogReader Stop() doesn't cause use-after-free
- *
- * Verifies that Stop() properly signals shutdown and that reader thread
- * checks should_stop_ after returning from blocking mysql_binlog_fetch().
- *
- * NOTE: This is a structural/lifecycle test. The actual fix (checking should_stop_
- * after mysql_binlog_fetch returns) is verified in integration tests with real
- * MySQL connections, as unit tests cannot easily simulate the blocking call.
- */
-TEST(BinlogReaderTest, StopDoesNotCauseUseAfterFree) {
-  Connection::Config config;
-  config.host = "localhost";
-  config.user = "test";
-  config.database = "test";
-
-  Connection conn(config);
-
-  // Create table contexts (not actually used in this test)
-  std::unordered_map<std::string, server::TableContext*> table_contexts;
-
-  BinlogReader::Config reader_config;
-  BinlogReader reader(conn, table_contexts, reader_config);
-
-  // Start and immediately stop
-  // (Start will fail without real connection, but that's ok for this test)
-  reader.Stop();
-
-  // Verify Stop() completes without hanging
-  // The fix ensures that should_stop_ is checked after mysql_binlog_fetch() returns,
-  // preventing use-after-free when connection is closed during Stop()
-  SUCCEED();
-}
-
-/**
- * @brief Test reconnection delay reset behavior
- *
- * Verifies that reconnection attempt counter is properly managed during
- * connection failures and successful reconnections.
- *
- * NOTE: This is a documentation test. The actual behavior (resetting reconnect_attempt
- * to 0 after successful reconnection) is verified in integration tests with real
- * MySQL connections. The fix prevents infinite delay increase by resetting the
- * counter when reconnection succeeds.
- *
- * Key behaviors tested in integration tests:
- * 1. reconnect_attempt increments on connection failure
- * 2. Delay increases exponentially: delay = base_delay * min(attempt, 10)
- * 3. reconnect_attempt resets to 0 after successful reconnection
- * 4. Prevents unbounded delay growth in long-running systems
- */
-TEST(BinlogReaderTest, ReconnectionDelayResetBehaviorDocumented) {
-  // This test documents the expected behavior for reconnection delay management
-  //
-  // Before fix:
-  //   - reconnect_attempt never reset after successful reconnection
-  //   - Delay would stay at maximum (10x base delay) forever
-  //   - Long-running systems would have unnecessarily long reconnection delays
-  //
-  // After fix:
-  //   - reconnect_attempt resets to 0 after any successful reconnection
-  //   - Subsequent failures start from base delay again
-  //   - Better recovery behavior for transient connection issues
-  //
-  // The actual verification requires integration tests with MySQL connection
-  // failures and recovery scenarios.
-
-  SUCCEED();
-}
-
-/**
- * @brief Test exception handling in filter value parsing
- * Regression test for: std::stod() and std::stoull() had no exception handling
- */
-TEST(BinlogReaderFilterTest, InvalidFilterValueExceptionHandling) {
-  // Test invalid float value (stod exception)
-  config::RequiredFilterConfig float_filter;
-  float_filter.name = "score";
-  float_filter.type = "double";
-  float_filter.op = "=";
-  float_filter.value = "not_a_number";  // Invalid float
-
-  storage::FilterValue test_value = 3.14;
-
-  // Should not crash, should return false
-  bool result = BinlogFilterEvaluator::CompareFilterValue(test_value, float_filter);
-  EXPECT_FALSE(result);
-
-  // Test invalid datetime value (stoull exception)
-  config::RequiredFilterConfig datetime_filter;
-  datetime_filter.name = "created_at";
-  datetime_filter.type = "unsigned";
-  datetime_filter.op = "=";
-  datetime_filter.value = "invalid_timestamp";  // Invalid uint64
-
-  storage::FilterValue datetime_value = uint64_t(1234567890);
-
-  // Should not crash, should return false
-  result = BinlogFilterEvaluator::CompareFilterValue(datetime_value, datetime_filter);
-  EXPECT_FALSE(result);
-
-  // Test overflow in stoull
-  datetime_filter.value = "99999999999999999999999999";  // Way too large
-  result = BinlogFilterEvaluator::CompareFilterValue(datetime_value, datetime_filter);
-  EXPECT_FALSE(result);
-
-  // Test valid values for comparison
-  float_filter.value = "3.14";
-  result = BinlogFilterEvaluator::CompareFilterValue(test_value, float_filter);
-  EXPECT_TRUE(result);
-
-  datetime_filter.value = "1234567890";
-  result = BinlogFilterEvaluator::CompareFilterValue(datetime_value, datetime_filter);
-  EXPECT_TRUE(result);
-}
-
-/**
- * @brief Test multi-table mode with different table configurations
- *
- * Regression test for: Multi-table mode was using global table_config_ instead of
- * per-table configuration, causing incorrect text_column, primary_key, and filter extraction.
- *
- * This test ensures each table uses its own configuration independently.
- */
-TEST_F(BinlogReaderFixture, MultiTableModeUsesCorrectTableConfig) {
-  // Create articles table with "content" as text column
-  server::TableContext articles_ctx;
-  articles_ctx.name = "articles";
-  articles_ctx.config.name = "articles";
-  articles_ctx.config.primary_key = "article_id";
-  articles_ctx.config.text_source.column = "content";
-  articles_ctx.index = std::make_unique<index::Index>(2);
-  articles_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  config::FilterConfig article_filter;
-  article_filter.name = "author_id";
-  article_filter.type = "int";
-  articles_ctx.config.filters.push_back(article_filter);
-
-  // Create comments table with DIFFERENT configuration
-  server::TableContext comments_ctx;
-  comments_ctx.name = "comments";
-  comments_ctx.config.name = "comments";
-  comments_ctx.config.primary_key = "comment_id";   // Different primary key
-  comments_ctx.config.text_source.column = "body";  // Different text column
-  comments_ctx.index = std::make_unique<index::Index>(2);
-  comments_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  config::FilterConfig comment_filter;
-  comment_filter.name = "post_id";  // Different filter
-  comment_filter.type = "int";
-  comments_ctx.config.filters.push_back(comment_filter);
-
-  std::unordered_map<std::string, server::TableContext*> contexts = {
-      {articles_ctx.name, &articles_ctx},
-      {comments_ctx.name, &comments_ctx},
-  };
-
-  BinlogReader multi_reader(connection_, contexts, reader_config_);
-
-  // Test ExtractAllFilters with articles config
-  RowData article_row;
-  article_row.primary_key = "100";
-  article_row.text = "Article text";
-  article_row.columns["author_id"] = "42";  // RowData.columns is string-to-string map
-
-  auto article_filters = BinlogFilterEvaluator::ExtractAllFilters(article_row, articles_ctx.config);
-  // Verify articles table extracts author_id (not post_id)
-  EXPECT_TRUE(article_filters.find("author_id") != article_filters.end());
-  // Articles should NOT have post_id since it's not in the config
-  EXPECT_TRUE(article_filters.find("post_id") == article_filters.end());
-
-  // Test ExtractAllFilters with comments config - should NOT extract author_id
-  RowData comment_row;
-  comment_row.primary_key = "200";
-  comment_row.text = "Comment text";
-  comment_row.columns["post_id"] = "999";   // RowData.columns is string-to-string map
-  comment_row.columns["author_id"] = "42";  // Also add author_id to row data
-
-  auto comment_filters = BinlogFilterEvaluator::ExtractAllFilters(comment_row, comments_ctx.config);
-  // Verify comments table extracts post_id (not author_id)
-  EXPECT_TRUE(comment_filters.find("post_id") != comment_filters.end());
-  // Comments should NOT have author_id since it's not in the config
-  EXPECT_TRUE(comment_filters.find("author_id") == comment_filters.end());
-}
-
-/**
- * @brief Test multi-table mode with different required filters
- *
- * Ensures that required_filters from each table's config are correctly applied,
- * not mixing up between tables.
- */
-TEST_F(BinlogReaderFixture, MultiTableModeRequiredFiltersPerTable) {
-  // Table 1: Accepts status = 1
-  server::TableContext published_ctx;
-  published_ctx.name = "published";
-  published_ctx.config.name = "published";
-  published_ctx.config.primary_key = "id";
-  published_ctx.config.text_source.column = "content";
-  published_ctx.index = std::make_unique<index::Index>(2);
-  published_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  config::RequiredFilterConfig published_filter;
-  published_filter.name = "status";
-  published_filter.type = "int";
-  published_filter.op = "=";
-  published_filter.value = "1";
-  published_ctx.config.required_filters.push_back(published_filter);
-
-  // Table 2: Accepts status = 0
-  server::TableContext draft_ctx;
-  draft_ctx.name = "drafts";
-  draft_ctx.config.name = "drafts";
-  draft_ctx.config.primary_key = "id";
-  draft_ctx.config.text_source.column = "content";
-  draft_ctx.index = std::make_unique<index::Index>(2);
-  draft_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  config::RequiredFilterConfig draft_filter;
-  draft_filter.name = "status";
-  draft_filter.type = "int";
-  draft_filter.op = "=";
-  draft_filter.value = "0";
-  draft_ctx.config.required_filters.push_back(draft_filter);
-
-  std::unordered_map<std::string, server::TableContext*> contexts = {
-      {published_ctx.name, &published_ctx},
-      {draft_ctx.name, &draft_ctx},
-  };
-
-  BinlogReader multi_reader(connection_, contexts, reader_config_);
-
-  // Filters with status = 1
-  std::unordered_map<std::string, storage::FilterValue> filters_published;
-  filters_published["status"] = static_cast<int64_t>(1);
-
-  // Filters with status = 0
-  std::unordered_map<std::string, storage::FilterValue> filters_draft;
-  filters_draft["status"] = static_cast<int64_t>(0);
-
-  // Published table should accept status=1, reject status=0
-  EXPECT_TRUE(BinlogFilterEvaluator::EvaluateRequiredFilters(filters_published, published_ctx.config, "+00:00"));
-  EXPECT_FALSE(BinlogFilterEvaluator::EvaluateRequiredFilters(filters_draft, published_ctx.config, "+00:00"));
-
-  // Draft table should accept status=0, reject status=1
-  EXPECT_FALSE(BinlogFilterEvaluator::EvaluateRequiredFilters(filters_published, draft_ctx.config, "+00:00"));
-  EXPECT_TRUE(BinlogFilterEvaluator::EvaluateRequiredFilters(filters_draft, draft_ctx.config, "+00:00"));
-}
-
-/**
- * @brief Test multi-table mode with concat vs single column text source
- *
- * Verifies that tables with different text_source configurations work correctly
- * in multi-table mode (one using single column, another using concat).
- */
-TEST_F(BinlogReaderFixture, MultiTableModeDifferentTextSources) {
-  // Table 1: Single column text source
-  server::TableContext products_ctx;
-  products_ctx.name = "products";
-  products_ctx.config.name = "products";
-  products_ctx.config.primary_key = "id";
-  products_ctx.config.text_source.column = "name";  // Single column
-  products_ctx.config.text_source.concat.clear();
-  products_ctx.index = std::make_unique<index::Index>(2);
-  products_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  // Table 2: Concat text source (multiple columns)
-  server::TableContext users_ctx;
-  users_ctx.name = "users";
-  users_ctx.config.name = "users";
-  users_ctx.config.primary_key = "user_id";
-  users_ctx.config.text_source.column.clear();  // No single column
-  users_ctx.config.text_source.concat = {"first_name", "last_name", "email"};
-  users_ctx.index = std::make_unique<index::Index>(2);
-  users_ctx.doc_store = std::make_unique<storage::DocumentStore>();
-
-  std::unordered_map<std::string, server::TableContext*> contexts = {
-      {products_ctx.name, &products_ctx},
-      {users_ctx.name, &users_ctx},
-  };
-
-  BinlogReader multi_reader(connection_, contexts, reader_config_);
-
-  // Verify each table has correct configuration
-  EXPECT_FALSE(products_ctx.config.text_source.column.empty());
-  EXPECT_TRUE(products_ctx.config.text_source.concat.empty());
-
-  EXPECT_TRUE(users_ctx.config.text_source.column.empty());
-  EXPECT_FALSE(users_ctx.config.text_source.concat.empty());
-  EXPECT_EQ(users_ctx.config.text_source.concat.size(), 3);
-}
-
-/**
- * @brief Test null pointer safety in table context handling
- * Regression test for: table_iter->second->index/doc_store could be null
- *
- * Note: This is a documentation test since creating a TableContext with null
- * index/doc_store in production code is prevented by design. The actual fix
- * adds defensive null checks to prevent crashes if this ever happens.
- *
- * The modified code path is:
- * - src/mysql/binlog_reader.cpp:665-668
- *
- * Now checks for null pointers and logs error instead of crashing.
- */
-TEST(BinlogReaderPointerSafetyTest, NullTableContextDefensiveChecks) {
-  // This test documents the safety improvements
-  // In practice, the null pointer checks in binlog_reader.cpp prevent crashes when:
-  // 1. A table is registered but index/doc_store initialization fails
-  // 2. A table is in an inconsistent state during reconfiguration
-  // 3. Memory corruption or other unexpected conditions occur
-
-  // The fix ensures:
-  // - No segfault/crash occurs
-  // - Error is logged
-  // - Binlog event processing fails gracefully
-  // - Reader continues with next event
-
-  SUCCEED() << "Null pointer safety checks added to binlog event processing";
-}
-
-/**
- * @brief Test concurrent Start() calls don't cause race conditions
- * Regression test for: gtid_encoded_data_ was not protected by mutex
- * and running_ flag was not atomically checked-and-set
- */
-TEST_F(BinlogReaderFixture, ConcurrentStartCallsThreadSafe) {
-  // Attempt to start the reader from multiple threads concurrently
-  std::atomic<int> start_success_count{0};
-  std::atomic<int> already_running_count{0};
-
-  constexpr int num_threads = 5;
-  std::vector<std::thread> threads;
-
-  for (int i = 0; i < num_threads; ++i) {
-    threads.emplace_back([&]() {
-      // Note: Start() will fail because we don't have a real MySQL connection
-      // But the important part is testing the thread safety of the check-and-set logic
-      bool result = reader_->Start();
-      if (result) {
-        start_success_count.fetch_add(1);
-      } else {
-        // Check if error was "already running"
-        if (reader_->GetLastError().find("already running") != std::string::npos) {
-          already_running_count.fetch_add(1);
-        }
-      }
-    });
-  }
-
-  for (auto& thread : threads) {
-    thread.join();
-  }
-
-  // With proper atomic compare_exchange_strong:
-  // - At most one thread should succeed or all should fail for other reasons
-  // - Multiple threads should get "already running" error
-  // The key is: no race condition or crash should occur
-  EXPECT_LE(start_success_count.load(), 1) << "At most one Start() should succeed";
-
-  // Stop the reader if it was started
-  if (reader_->IsRunning()) {
-    reader_->Stop();
-  }
-}
-
-/**
- * @brief Test exponential backoff cap behavior
- * Regression test for: reconnect_attempt could grow unbounded
- * Verifies that reconnect delay is properly capped at 10x base delay
- */
-TEST(BinlogReaderTest, ExponentialBackoffCapped) {
-  // This test documents the expected behavior for reconnection backoff
-
-  // The implementation should ensure:
-  // - reconnect_attempt is capped at 10 using std::min(reconnect_attempt + 1, 10)
-  // - This prevents integer overflow and unbounded delays
-  // - Maximum delay is base_delay * 10
-
-  // Example with base delay of 1000ms:
-  // Attempt 1: 1000ms * 1 = 1000ms
-  // Attempt 2: 1000ms * 2 = 2000ms
-  // ...
-  // Attempt 10: 1000ms * 10 = 10000ms
-  // Attempt 11+: 1000ms * 10 = 10000ms (capped)
-
-  // Without the cap:
-  // - After many reconnection attempts, delay_ms could overflow
-  // - System would have excessive delays during network issues
-
-  // With the cap (current implementation):
-  // - Maximum delay is predictable: config.reconnect_delay_ms * 10
-  // - No overflow risk
-  // - Reasonable maximum backoff time
-
-  SUCCEED() << "Exponential backoff is capped at 10x base delay (src/mysql/binlog_reader.cpp:380)";
-}
-
-/**
- * @brief Test multi-table DDL processing
- * Regression test for: QUERY_EVENT only checked single table_config_.name
- * Verifies that DDL events are properly detected for all registered tables
- */
-TEST(BinlogReaderTest, MultiTableDDLProcessing) {
-  // This test documents the expected behavior for multi-table DDL handling
-
-  // In multi-table mode (multi_table_mode_ == true):
-  // - QUERY_EVENT (DDL) should check all tables in table_contexts_
-  // - DDL affecting any registered table should be detected
-  // - Example: "ALTER TABLE table1 ..." should be caught if table1 is registered
-
-  // In single-table mode (multi_table_mode_ == false):
-  // - QUERY_EVENT should only check table_config_.name
-  // - Only DDL affecting the configured table is processed
-
-  // The implementation (src/mysql/binlog_reader.cpp:1181-1201):
-  // if (multi_table_mode_) {
-  //   for (const auto& [table_name, ctx] : table_contexts_) {
-  //     if (IsTableAffectingDDL(query, table_name)) {
-  //       // Process DDL for this table
-  //     }
-  //   }
-  // } else {
-  //   if (IsTableAffectingDDL(query, table_config_.name)) {
-  //     // Process DDL for single table
-  //   }
-  // }
-
-  // Without this fix:
-  // - Multi-table mode would only check table_config_.name
-  // - DDL for other registered tables would be missed
-  // - Schema changes could be lost
-
-  // With this fix:
-  // - All registered tables are checked for DDL
-  // - Proper multi-table DDL handling
-
-  SUCCEED()
-      << "Multi-table DDL processing properly iterates all registered tables (src/mysql/binlog_reader.cpp:1181-1201)";
-}
-
-/**
- * @brief Test filter value size validation (security: memory exhaustion protection)
- */
-TEST_F(BinlogReaderFixture, FilterValueSizeValidation) {
-  // Normal size filter value should work
-  config::RequiredFilterConfig normal_filter;
-  normal_filter.name = "status";
-  normal_filter.op = "=";
-  normal_filter.value = "active";  // Small value
-
-  storage::FilterValue test_value = std::string("active");
-  EXPECT_TRUE(BinlogFilterEvaluator::CompareFilterValue(test_value, normal_filter))
-      << "Normal-sized filter value should be accepted";
-
-  // Large but acceptable filter value (< 1MB)
-  config::RequiredFilterConfig large_filter;
-  large_filter.name = "description";
-  large_filter.op = "=";
-  large_filter.value = std::string(100 * 1024, 'x');  // 100KB
-
-  storage::FilterValue large_test_value = large_filter.value;
-  EXPECT_TRUE(BinlogFilterEvaluator::CompareFilterValue(large_test_value, large_filter))
-      << "Large filter value (100KB) should be accepted";
-
-  // Oversized filter value (> 1MB) should be rejected
-  config::RequiredFilterConfig oversized_filter;
-  oversized_filter.name = "malicious";
-  oversized_filter.op = "=";
-  oversized_filter.value = std::string(2 * 1024 * 1024, 'x');  // 2MB (exceeds limit)
-
-  storage::FilterValue oversized_test_value = std::string("test");
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(oversized_test_value, oversized_filter))
-      << "Oversized filter value (2MB) should be rejected for security";
-
-  // Edge case: exactly at limit (1MB)
-  config::RequiredFilterConfig edge_filter;
-  edge_filter.name = "edge_case";
-  edge_filter.op = "=";
-  edge_filter.value = std::string(1024 * 1024, 'y');  // Exactly 1MB
-
-  storage::FilterValue edge_test_value = edge_filter.value;
-  EXPECT_TRUE(BinlogFilterEvaluator::CompareFilterValue(edge_test_value, edge_filter))
-      << "Filter value at exact limit (1MB) should be accepted";
-
-  // Just over limit (1MB + 1 byte)
-  config::RequiredFilterConfig just_over_filter;
-  just_over_filter.name = "just_over";
-  just_over_filter.op = "=";
-  just_over_filter.value = std::string(1024 * 1024 + 1, 'z');  // 1MB + 1 byte
-
-  storage::FilterValue just_over_test_value = std::string("test");
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(just_over_test_value, just_over_filter))
-      << "Filter value just over limit (1MB+1) should be rejected";
-}
-
-/**
- * @brief Test filter value size validation with different data types
- */
-TEST_F(BinlogReaderFixture, FilterValueSizeValidationTypes) {
-  // Integer filter with oversized string representation
-  config::RequiredFilterConfig int_filter;
-  int_filter.name = "number";
-  int_filter.op = "=";
-  int_filter.value = std::string(2 * 1024 * 1024, '9');  // 2MB of '9's
-
-  storage::FilterValue int_value = int64_t(999);
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(int_value, int_filter))
-      << "Oversized integer filter value string should be rejected";
-
-  // Double filter with oversized string representation
-  config::RequiredFilterConfig double_filter;
-  double_filter.name = "price";
-  double_filter.op = "=";
-  double_filter.value = std::string(2 * 1024 * 1024, '1');  // 2MB
-
-  storage::FilterValue double_value = 123.45;
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(double_value, double_filter))
-      << "Oversized double filter value string should be rejected";
-
-  // Datetime filter with oversized string representation
-  config::RequiredFilterConfig datetime_filter;
-  datetime_filter.name = "created_at";
-  datetime_filter.op = "=";
-  datetime_filter.value = std::string(2 * 1024 * 1024, '2');  // 2MB
-
-  storage::FilterValue datetime_value = uint64_t(1234567890);
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(datetime_value, datetime_filter))
-      << "Oversized datetime filter value string should be rejected";
-}
-
-/**
- * @brief Test that NULL checks work regardless of filter value size
- */
-TEST_F(BinlogReaderFixture, FilterValueSizeValidationNullChecks) {
-  // IS NULL should work even with oversized filter value
-  config::RequiredFilterConfig null_filter;
-  null_filter.name = "deleted_at";
-  null_filter.op = "IS NULL";
-  null_filter.value = std::string(2 * 1024 * 1024, 'x');  // Oversized (but ignored for IS NULL)
-
-  storage::FilterValue null_value = std::monostate{};
-  // IS NULL doesn't use filter.value, so size check happens but doesn't affect NULL check
-  // The function returns false early due to size check before reaching NULL logic
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(null_value, null_filter))
-      << "Oversized filter value should be rejected even for NULL checks";
-
-  // IS NOT NULL should also be affected by size validation
-  config::RequiredFilterConfig not_null_filter;
-  not_null_filter.name = "updated_at";
-  not_null_filter.op = "IS NOT NULL";
-  not_null_filter.value = std::string(2 * 1024 * 1024, 'y');  // Oversized
-
-  storage::FilterValue non_null_value = uint64_t(1234567890);
-  EXPECT_FALSE(BinlogFilterEvaluator::CompareFilterValue(non_null_value, not_null_filter))
-      << "Oversized filter value should be rejected even for NOT NULL checks";
-}
-
-#endif  // USE_MYSQL
